Se agregan gambeObjects con sprites para marcar los puntos de inicio y final. El codigo es correcto pero por algun motivo no encuentra las imagenes.
