--- Source/SpawnPoint.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using ICities;
-using UnityEngine;
-using ColossalFramework.Plugins;
-using ColossalFramework.UI;
-
-namespace ODStudy
-{
-
-}
